Organize test files
